+++ core/common/src/thrift/block_master.thrift
@@ -4,8 +4,8 @@ include "common.thrift"
 include "exception.thrift"
 
 struct BlockMasterInfo {
- 1: i64 liveWorkerNum
- 2: i64 deadWorkerNum
+ 1: i32 liveWorkerNum
+ 2: i32 deadWorkerNum
  3: string TotalCapacity
  4: string usedCapacity
  5: string FreeCapacity
